+++ lib/predicator/conformance/coverage.ex
@@ -140,7 +140,8 @@ defmodule Predicator.Conformance.Coverage do
     5 => "functions",
     6 => "statements",
     7 => "casts",
-    8 => "control flow"
+    8 => "control flow",
+    9 => "loops"
   }
 
   @doc """